+++ lib/splendid_cli.dart
@@ -34,6 +34,7 @@ import 'src/commands/format_command.dart';
 import 'src/commands/gui_command.dart';
 import 'src/commands/screen_command.dart';
 import 'src/commands/setup_command.dart';
+import 'src/commands/sort_l10n_command.dart';
 import 'src/commands/test_command.dart';
 import 'src/utils/custom_help.dart';
 
@@ -43,6 +44,7 @@ export 'src/mcp/mcp_tool_registry.dart';
 
 // Export services for external use
 export 'src/services/dartdoc_formatter_service.dart';
+export 'src/services/l10n_sorter_service.dart';
 export 'src/services/project_service.dart';
 export 'src/services/screen_service.dart';
 export 'src/services/test_service.dart';
@@ -72,6 +74,7 @@ class SplendidCommandRunner extends CommandRunner<int> {
     addCommand(GuiCommand());
     addCommand(ScreenCommand());
     addCommand(SetupCommand());
+    addCommand(SortL10nCommand());
     addCommand(TestCommand());
   }